Delete Authorization

+++ App/controllers/ListingController.php
@@ -4,6 +4,8 @@ namespace App\Controllers;
 
 use Framework\Database;
 use Framework\Validation;
+use Framework\Session;
+use Framework\Authorization;
 
 class ListingController {
     protected $db;
@@ -13,11 +15,12 @@ class ListingController {
         $this->db = new Database($config);
     }
 
-    public function index() {
-        $listings = $this->db->query('SELECT * FROM listings')->fetchAll();
+    public function index(){
+        $listings = $this->db->query('SELECT * FROM listings ORDER BY created_at DESC')->fetchAll();
+
 
         loadView('listings/index', [
-            'listings' => $listings,
+            'listings' => $listings
         ]);
     }
 
@@ -63,7 +66,7 @@ class ListingController {
 
         $newListingData = array_intersect_key($_POST, array_flip($allowedFields));
 
-        $newListingData['user_id'] = 1;
+        $newListingData['user_id'] = Session::get('user')['id'];
 
         $newListingData = array_map('sanitize', $newListingData);
 
@@ -126,13 +129,21 @@ class ListingController {
         ];
         $listing = $this->db->query('SELECT * FROM listings WHERE id = :id', $params)->fetch();
 
+        // Check if listing exists
         if(!$listing) {
             ErrorController::notFound('Listing not found');
             return;
         }
 
+        // Authorization
+        if (!Authorization::isOwner($listing->user_id)) {
+            $_SESSION['error_message'] = 'You do not have authorized to deleted this listing';
+            return redirect('/listings/' . $listing->id);
+        }
+
         $this->db->query('DELETE FROM listings WHERE id = :id', $params);
 
+        // Set flash message
         $_SESSION['success_message'] = 'Listing deleted successfully';
 
         redirect('/listings');
